implement private route and login page using the createBrowserRouter
