+++ client/src/components/Layout/PrivateRoute.tsx
@@ -1,9 +1,10 @@
 import React, { ReactElement } from 'react';
 import { Navigate } from 'react-router-dom';
+import { useAppSelector } from 'store';
+import { selectUserName } from 'store/userSlice';
 
 interface Props {
   children: any;
-  authUser: string | undefined;
 }
 
 /**
@@ -11,7 +12,8 @@ interface Props {
  * @param { children } Route to wrap around
  * @constructor
  */
-export function PrivateRoute({ children, authUser }: Props): ReactElement {
+export function PrivateRoute({ children }: Props): ReactElement {
+  const authUser = useAppSelector(selectUserName);
   if (!authUser) {
     // not logged in so redirect to login page with the return url
     return <Navigate to="/login" />;
