--- client/src/components/PrivateRoute/index.ts
@@ -1,3 +1,0 @@
-import { PrivateRoute } from './PrivateRoute';
-
-export { PrivateRoute };
